requirements.txt
api==0.0.7
asgiref==3.8.1
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.3.2
decouple==0.0.7
Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
nose==1.3.7
psycopg2-binary==2.9.9
PyJWT==2.9.0
redis==5.1.0
requests==2.32.3
routing==0.2.0
sqlparse==0.5.1
urllib3==2.2.3
